Reduced clutter from within Hotel rooms and search command classes by extracting their nested types to the CAPI namespace.
