+++ Connexions.Travel/Capi/Hotel/RoomSearchInitRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace Connexions.Travel.Capi.Hotel
+{
+	class RoomSearchInitRequest : SearchInitRequest
+	{
+		public string hotelId;
+	}
+}
