+++ Connexions.Travel/Capi/Hotel/RoomSearchStatusResponse.cs
@@ -1,0 +1,10 @@
+﻿namespace Connexions.Travel.Capi.Hotel
+{
+	class RoomSearchStatusResponse : StatusResponse
+	{
+		/// <summary>
+		/// Total count of hotel results so far.
+		/// </summary>
+		public int roomCount;
+	}
+}
