+++ Connexions.Travel/Capi/Hotel/SearchStatusResponse.cs
@@ -1,0 +1,10 @@
+﻿namespace Connexions.Travel.Capi.Hotel
+{
+	class SearchStatusResponse : StatusResponse
+	{
+		/// <summary>
+		/// Total count of hotel results so far.
+		/// </summary>
+		public int hotelCount;
+	}
+}
